PasswordInput enabling addons in Django 1.11
At least in Django 1.11 (not sure in which version this change was introduced), PasswordInput is no longe a subclass of TextInput and is therefore nor recognised in check in the make_input_group method. This prevents the addon_before and addon_after to work on these fields.
This is fixed in this commit.